+++ Stripe/STPCard+Private.h
@@ -16,7 +16,6 @@ NS_ASSUME_NONNULL_BEGIN
 + (STPCardFundingType)fundingFromString:(NSString *)string;
 + (nullable NSString *)stringFromFunding:(STPCardFundingType)funding;
 
-+ (STPCardBrand)brandFromString:(NSString *)string;
 + (NSString *)stringFromBrand:(STPCardBrand)brand;
 
 @end