+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 11.x.x 2017-XX-XX
+
+* Restores `[STPCard brandFromString:]` method which was marked as deprecated in a recent version [#801](https://github.com/stripe/stripe-ios/pull/801)
+
 ## 11.3.0 2017-09-13
 * Adds support for creating `STPSourceParams` for P24 source [#779](https://github.com/stripe/stripe-ios/pull/779)
 * Adds support for native app-to-app Alipay redirects [#783](https://github.com/stripe/stripe-ios/pull/783)